# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DocScan — Document Scan & Email System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/{id}/send:
    post:
      operationId: sendDocument
      tags: [documents]
      summary: Send document to configured recipients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SendResult:
      type: object
      required: [success, logs]
      properties:
        success:
          type: boolean
        message:
          type: string
        logs:
          type: array
          items:
            $ref: "#/components/schemas/EmailLog"
    EmailLog:
      type: object
      required: [id, documentId, senderId, recipientEmail, status, sentAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        senderId:
          type: integer
        recipientEmail:
          type: string
        status:
          type: string
          enum: [queued, sent, failed]
        sentAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        documentName:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
...