# lib/api-zod/src/generated/types/emailLog.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DocScan — Document Scan & Email System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailLogStatus } from './emailLogStatus';

export interface EmailLog {
  id: number;
  documentId: number;
  senderId: number;
  recipientEmail: string;
  status: EmailLogStatus;
  /** @nullable */
  sentAt: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  documentName?: string | null;
  /** @nullable */
  senderName?: string | null;
}